# internal/issue/handler.go
// Package issue provides HTTP handlers for issue workflows.
package issue

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"log"
	"mime/multipart"
	"net/http"
	"os"
	"strconv"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"

	"6s/internal/apperror"
	"6s/internal/auth"
	"6s/internal/db"
	"6s/internal/i18n"
	"6s/internal/response"
)

// Service defines operations required by HTTP handlers.
type Service interface {
	SyncIssue(ctx context.Context, req SyncIssueRequest, currentUser db.User) (*Response, bool, error)
	ResolveIssue(ctx context.Context, req ResolveIssueRequest, currentUser db.User) (*Response, error)
	CloseIssue(ctx context.Context, req CloseIssueRequest, currentUser db.User) (*Response, error)
	ReopenIssue(ctx context.Context, req ReopenIssueRequest, currentUser db.User) (*Response, error)
	InvalidateIssue(ctx context.Context, req InvalidateIssueRequest, currentUser db.User) (*Response, error)
	PatchIssue(ctx context.Context, req PatchIssueRequest, currentUser db.User) (*Response, error)
	GetIssueByID(ctx context.Context, id int64) (*Response, error)
	OpenMedia(ctx context.Context, id int64, folder, basename string) (*os.File, error)
	ListIssuesFiltered(ctx context.Context, filter ListFilter) ([]Response, int64, error)
	SubscribeEvents(userID ...int64) (<-chan Event, func())
}

// Handler handles Issue HTTP endpoints.
type Handler struct {
	service Service
}

// NewHandler creates a new Issue Handler.
func NewHandler(service Service) *Handler {
	return &Handler{service: service}
}

func parseQueryValues(q map[string][]string, singularKey, pluralKey string) []string {
	var results []string
	for _, key := range []string{pluralKey, singularKey} {
		for _, val := range q[key] {
			for _, part := range strings.Split(val, ",") {
				if s := strings.TrimSpace(part); s != "" {
					results = append(results, s)
				}
			}
		}
	}
	return results
}

// List handles GET /api/issues.
func (h *Handler) List(w http.ResponseWriter, r *http.Request) {
	q := r.URL.Query()
	statuses := parseQueryValues(q, "status", "statuses")
	categories := parseQueryValues(q, "category", "categories")
	locationCodes := parseQueryValues(q, "location_code", "location_codes")
	var tagCode *string
	if raw := strings.TrimSpace(q.Get("tag_code")); raw != "" {
		tagCode = &raw
	}
	overdue := q.Get("overdue") == "true" || q.Get("overdue") == "1"
	page, limit := response.ParsePageLimit(q, 20, 100)
	filter := ListFilter{Statuses: statuses, Categories: categories, LocationCodes: locationCodes, TagCode: tagCode, Overdue: overdue, Page: page, Limit: limit}
	if raw := q.Get("assigned_team_id"); raw != "" {
		id, err := strconv.ParseInt(raw, 10, 64)
		if err != nil || id <= 0 {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
			return
		}
		filter.AssignedTeamID = &id
	}
	filter.MineTeam = q.Get("mine_team") == "true" || q.Get("mine_team") == "1"
	items, total, err := h.service.ListIssuesFiltered(r.Context(), filter)
	if err != nil {
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrIssueListFailed).WithCause(err))
		return
	}
	_ = response.Paginated(w, http.StatusOK, items, page, limit, int(total))
}

// GetByID handles GET /api/issues/{id}.
func (h *Handler) GetByID(w http.ResponseWriter, r *http.Request) {
	id, err := response.ParseIDParam(r, "id")
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(err))
		return
	}

	resp, err := h.service.GetIssueByID(r.Context(), id)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		_ = response.AppError(w, r, apperror.Internal(i18n.ErrIssueGetFailed).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, resp)
}

// Media handles GET /api/issues/{id}/media/{folder}/{filename}.
func (h *Handler) Media(w http.ResponseWriter, r *http.Request) {
	id, err := response.ParseIDParam(r, "id")
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(err))
		return
	}
	folder := chi.URLParam(r, "folder")
	filename := chi.URLParam(r, "filename")

	f, err := h.service.OpenMedia(r.Context(), id, folder, filename)
	if err != nil {
		_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
		return
	}
	defer func() {
		_ = f.Close()
	}()

	stat, err := f.Stat()
	if err != nil {
		_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
		return
	}
	w.Header().Set("Cache-Control", "private, no-cache")
	http.ServeContent(w, r, stat.Name(), stat.ModTime(), f)
}

// Sync handles POST /api/issues/sync (Multipart form).
func (h *Handler) Sync(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	if err := r.ParseMultipartForm(10 * 1024 * 1024); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMultipartTooLarge).WithCause(err))
		return
	}

	clientUUID := strings.TrimSpace(r.FormValue("client_uuid"))
	category := strings.TrimSpace(r.FormValue("category"))
	causeType := strings.TrimSpace(r.FormValue("cause_type"))
	locationCode := strings.TrimSpace(r.FormValue("location_code"))
	locationNameViSnapshot := strings.TrimSpace(r.FormValue("location_name_vi_snapshot"))
	locationNameZhSnapshot := strings.TrimSpace(r.FormValue("location_name_zh_snapshot"))
	locationNameEnSnapshot := strings.TrimSpace(r.FormValue("location_name_en_snapshot"))
	locationSnapshotSource := strings.TrimSpace(r.FormValue("location_snapshot_source"))
	description := strings.TrimSpace(r.FormValue("description"))
	tagsStr := strings.TrimSpace(r.FormValue("tags"))
	proposedTagsStr := strings.TrimSpace(r.FormValue("proposed_tags"))

	if clientUUID == "" || category == "" || locationCode == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMissingIssueFields))
		return
	}

	var tags []string
	if tagsStr != "" {
		if err := json.Unmarshal([]byte(tagsStr), &tags); err != nil {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
			return
		}
	}
	var proposedTags []ProposedTag
	if proposedTagsStr != "" {
		if err := json.Unmarshal([]byte(proposedTagsStr), &proposedTags); err != nil {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
			return
		}
	}

	photoBefore, _, fileErr := r.FormFile("photo_before")
	if fileErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMissingPhotoBefore).WithCause(fileErr))
		return
	}
	if cErr := photoBefore.Close(); cErr != nil {
		log.Printf("failed to close photoBefore: %v", cErr)
	}

	var photoBeforeHeader, photoDetailHeader *multipart.FileHeader
	if fhs := r.MultipartForm.File["photo_before"]; len(fhs) > 0 {
		photoBeforeHeader = fhs[0]
	}
	if fhs := r.MultipartForm.File["photo_detail"]; len(fhs) > 0 {
		photoDetailHeader = fhs[0]
	}
	parseOptionalID := func(key string) (*int64, error) {
		raw := strings.TrimSpace(r.FormValue(key))
		if raw == "" {
			return nil, nil
		}
		v, err := strconv.ParseInt(raw, 10, 64)
		if err != nil || v <= 0 {
			return nil, fmt.Errorf("invalid %s", key)
		}
		return &v, nil
	}
	assetID, idErr := parseOptionalID("asset_id")
	if idErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(idErr))
		return
	}
	assignedTeamID, idErr := parseOptionalID("assigned_team_id")
	if idErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(idErr))
		return
	}
	assigneeID, idErr := parseOptionalID("assignee_id")
	if idErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(idErr))
		return
	}
	syncReq := SyncIssueRequest{
		ClientUUID: clientUUID, Category: category, CauseType: causeType, LocationCode: locationCode,
		LocationNameViSnapshot: locationNameViSnapshot, LocationNameZhSnapshot: locationNameZhSnapshot,
		LocationNameEnSnapshot: locationNameEnSnapshot, LocationSnapshotSource: locationSnapshotSource,
		Tags: tags, ProposedTags: proposedTags, Description: description, AssetID: assetID, AssignedTeamID: assignedTeamID,
		AssigneeID: assigneeID, PhotoBefore: photoBeforeHeader, PhotoDetail: photoDetailHeader,
	}
	resp, created, err := h.service.SyncIssue(r.Context(), syncReq, currentUser)
	if err != nil {
		if errors.Is(err, ErrInvalidCategory) {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidCategory))
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrIssueSaveFailed, err.Error()).WithCause(err))
		return
	}

	statusCode := http.StatusOK
	if created {
		statusCode = http.StatusCreated
	}
	_ = response.JSON(w, statusCode, resp)
}

// Resolve handles POST /api/issues/{id}/resolve (Multipart form).
func (h *Handler) Resolve(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	id, parseErr := response.ParseIDParam(r, "id")
	if parseErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(parseErr))
		return
	}

	if err := r.ParseMultipartForm(10 * 1024 * 1024); err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidMultipart).WithCause(err))
		return
	}

	resolvedUUID := strings.TrimSpace(r.FormValue("resolved_client_uuid"))
	expVerStr := strings.TrimSpace(r.FormValue("expected_version"))
	forceStr := strings.TrimSpace(r.FormValue("force"))
	force := (forceStr == "true" || forceStr == "1")

	if resolvedUUID == "" {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMissingResolvedUUID))
		return
	}

	expectedVersion := int32(1)
	if expVerStr != "" {
		if ev, evErr := strconv.ParseInt(expVerStr, 10, 32); evErr == nil {
			expectedVersion = int32(ev) //nolint:gosec
		}
	}

	photoAfter, _, fileErr := r.FormFile("photo_after")
	if fileErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMissingPhotoAfter).WithCause(fileErr))
		return
	}
	if err := photoAfter.Close(); err != nil {
		log.Printf("failed to close photoAfter: %v", err)
	}

	var photoAfterHeader *multipart.FileHeader
	if fhs := r.MultipartForm.File["photo_after"]; len(fhs) > 0 {
		photoAfterHeader = fhs[0]
	}

	resp, err := h.service.ResolveIssue(r.Context(), ResolveIssueRequest{
		IssueID:            id,
		ResolvedClientUUID: resolvedUUID,
		ExpectedVersion:    expectedVersion,
		Force:              force,
		PhotoAfter:         photoAfterHeader,
	}, currentUser)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		if errors.Is(err, ErrIssueConflict) {
			appErr := apperror.Conflict("ISSUE_CONFLICT", i18n.ErrIssueVersionChanged).WithDetails(map[string]any{
				"current_status":  resp,
				"current_version": expectedVersion,
			})
			_ = response.AppError(w, r, appErr)
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, resp)
}

// CloseRequest defines payload for closing an issue.
type CloseRequest struct {
	ScoreRating     int16  `json:"score_rating"`
	ExpectedVersion *int32 `json:"expected_version"`
}

// Close handles POST /api/issues/{id}/close.
func (h *Handler) Close(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	id, err := response.ParseIDParam(r, "id")
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(err))
		return
	}

	var req CloseRequest
	if r.Body != nil {
		if decErr := json.NewDecoder(r.Body).Decode(&req); decErr != nil && !errors.Is(decErr, errors.New("EOF")) {
			log.Printf("decode close body err: %v", decErr)
		}
	}

	resp, err := h.service.CloseIssue(r.Context(), CloseIssueRequest{
		IssueID:         id,
		ScoreRating:     req.ScoreRating,
		ExpectedVersion: req.ExpectedVersion,
	}, currentUser)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		if errors.Is(err, ErrPermissionDenied) {
			_ = response.AppError(w, r, apperror.Forbidden(i18n.ErrIssueCloseForbidden))
			return
		}
		if errors.Is(err, ErrIssueConflict) {
			_ = response.AppError(w, r, apperror.Conflict("ISSUE_CONFLICT", i18n.ErrIssueConflict).WithCause(err))
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, resp)
}

// ReopenRequest defines payload for reopening an issue.
type ReopenRequest struct {
	RejectReason    string `json:"reject_reason"`
	ExpectedVersion *int32 `json:"expected_version"`
}

// Reopen handles POST /api/issues/{id}/reopen.
func (h *Handler) Reopen(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	id, err := response.ParseIDParam(r, "id")
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(err))
		return
	}

	var req ReopenRequest
	if r.Body != nil {
		if decErr := json.NewDecoder(r.Body).Decode(&req); decErr != nil && !errors.Is(decErr, errors.New("EOF")) {
			log.Printf("decode reopen body err: %v", decErr)
		}
	}

	resp, err := h.service.ReopenIssue(r.Context(), ReopenIssueRequest{
		IssueID:         id,
		RejectReason:    req.RejectReason,
		ExpectedVersion: req.ExpectedVersion,
	}, currentUser)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		if errors.Is(err, ErrPermissionDenied) {
			_ = response.AppError(w, r, apperror.Forbidden(i18n.ErrIssueReopenForbidden))
			return
		}
		if errors.Is(err, ErrIssueConflict) {
			_ = response.AppError(w, r, apperror.Conflict("ISSUE_CONFLICT", i18n.ErrIssueConflict).WithCause(err))
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, resp)
}

// InvalidRequest defines payload for invalidating an issue.
type InvalidRequest struct {
	Reason          string `json:"reason"`
	ExpectedVersion *int32 `json:"expected_version"`
}

// Invalid handles POST /api/issues/{id}/invalid.
func (h *Handler) Invalid(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	id, err := response.ParseIDParam(r, "id")
	if err != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(err))
		return
	}

	var req InvalidRequest
	if r.Body != nil {
		if decErr := json.NewDecoder(r.Body).Decode(&req); decErr != nil && !errors.Is(decErr, errors.New("EOF")) {
			log.Printf("decode invalid body err: %v", decErr)
		}
	}

	resp, err := h.service.InvalidateIssue(r.Context(), InvalidateIssueRequest{
		IssueID:         id,
		Reason:          req.Reason,
		ExpectedVersion: req.ExpectedVersion,
	}, currentUser)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		if errors.Is(err, ErrPermissionDenied) {
			_ = response.AppError(w, r, apperror.Forbidden(i18n.ErrIssueInvalidForbidden))
			return
		}
		if errors.Is(err, ErrIssueConflict) {
			_ = response.AppError(w, r, apperror.Conflict("ISSUE_CONFLICT", i18n.ErrIssueConflict).WithCause(err))
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}

	_ = response.JSON(w, http.StatusOK, resp)
}

// PatchRequest defines payload for in-place quick editing of an issue.
type PatchRequest struct {
	Category        *string  `json:"category"`
	CauseType       *string  `json:"cause_type"`
	LocationCode    *string  `json:"location_code"`
	Description     *string  `json:"description"`
	Tags            []string `json:"tags"`
	AssetID         **int64  `json:"asset_id"`
	AssignedTeamID  **int64  `json:"assigned_team_id"`
	AssigneeID      **int64  `json:"assignee_id"`
	CauseTeamID     **int64  `json:"cause_team_id"`
	CauseStatus     **string `json:"cause_status"`
	ExpectedVersion *int32   `json:"expected_version"`
}

func (h *Handler) parseMultipartPatch(r *http.Request) (PatchIssueRequest, error) {
	var patchReq PatchIssueRequest
	if parseErr := r.ParseMultipartForm(10 * 1024 * 1024); parseErr != nil {
		return patchReq, parseErr
	}
	if cat := strings.TrimSpace(r.FormValue("category")); cat != "" {
		patchReq.Category = &cat
	}
	if cause := strings.TrimSpace(r.FormValue("cause_type")); cause != "" {
		patchReq.CauseType = &cause
	}
	if loc := strings.TrimSpace(r.FormValue("location_code")); loc != "" {
		patchReq.LocationCode = &loc
	}
	if desc := strings.TrimSpace(r.FormValue("description")); desc != "" || r.Form.Has("description") {
		patchReq.Description = &desc
	}
	if tagsStr := strings.TrimSpace(r.FormValue("tags")); tagsStr != "" {
		var tags []string
		if unmarshalErr := json.Unmarshal([]byte(tagsStr), &tags); unmarshalErr == nil {
			patchReq.Tags = tags
		}
	}
	if fhs := r.MultipartForm.File["photo_before"]; len(fhs) > 0 {
		patchReq.PhotoBefore = fhs[0]
	}
	if fhs := r.MultipartForm.File["photo_detail"]; len(fhs) > 0 {
		patchReq.PhotoDetail = fhs[0]
	}
	return patchReq, nil
}

func (h *Handler) parseJSONPatch(r *http.Request) (PatchIssueRequest, error) {
	var req PatchRequest
	if decErr := json.NewDecoder(r.Body).Decode(&req); decErr != nil {
		return PatchIssueRequest{}, decErr
	}
	return PatchIssueRequest{
		Category: req.Category, CauseType: req.CauseType, LocationCode: req.LocationCode,
		Description: req.Description, Tags: req.Tags, AssetID: req.AssetID,
		AssignedTeamID: req.AssignedTeamID, AssigneeID: req.AssigneeID,
		CauseTeamID: req.CauseTeamID, CauseStatus: req.CauseStatus, ExpectedVersion: req.ExpectedVersion,
	}, nil
}

// Patch handles PATCH /api/issues/{id} (JSON or multipart/form-data).
func (h *Handler) Patch(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrUnauthorized))
		return
	}

	id, parseIDErr := response.ParseIDParam(r, "id")
	if parseIDErr != nil {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrInvalidID).WithCause(parseIDErr))
		return
	}

	var patchReq PatchIssueRequest
	var parseErr error
	if strings.HasPrefix(r.Header.Get("Content-Type"), "multipart/form-data") {
		patchReq, parseErr = h.parseMultipartPatch(r)
		if parseErr != nil {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrMultipartTooLarge).WithCause(parseErr))
			return
		}
	} else {
		patchReq, parseErr = h.parseJSONPatch(r)
		if parseErr != nil {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(parseErr))
			return
		}
	}
	patchReq.IssueID = id
	resp, err := h.service.PatchIssue(r.Context(), patchReq, currentUser)
	if err != nil {
		if errors.Is(err, ErrIssueNotFound) {
			_ = response.AppError(w, r, apperror.NotFound(i18n.ErrIssueNotFound))
			return
		}
		if errors.Is(err, ErrPermissionDenied) {
			_ = response.AppError(w, r, apperror.Forbidden(i18n.ErrIssuePatchForbidden))
			return
		}
		if errors.Is(err, ErrIssueConflict) {
			_ = response.AppError(w, r, apperror.Conflict("ISSUE_CONFLICT", i18n.ErrIssueVersionChanged).WithCause(err))
			return
		}
		if errors.Is(err, ErrInvalidCategory) || errors.Is(err, ErrInvalidResponsibility) || errors.Is(err, ErrMissingExpectedVersion) {
			_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
			return
		}
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(err))
		return
	}
	_ = response.JSON(w, http.StatusOK, resp)
}

// Events streams real-time issue updates via Server-Sent Events (SSE).
func (h *Handler) Events(w http.ResponseWriter, r *http.Request) {
	currentUser, ok := auth.GetUserFromContext(r.Context())
	if !ok {
		_ = response.AppError(w, r, apperror.Unauthorized(i18n.ErrMissingAuth))
		return
	}
	flusher, ok := w.(http.Flusher)
	if !ok {
		_ = response.AppError(w, r, apperror.BadRequest(i18n.ErrBadRequest).WithCause(errors.New("streaming unsupported")))
		return
	}

	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Set("Cache-Control", "no-cache, no-transform")
	w.Header().Set("Connection", "keep-alive")
	w.Header().Set("X-Accel-Buffering", "no")
	w.WriteHeader(http.StatusOK)
	flusher.Flush()

	eventsCh, unsubscribe := h.service.SubscribeEvents(currentUser.ID)
	defer unsubscribe()

	ticker := time.NewTicker(25 * time.Second)
	defer ticker.Stop()

	for {
		select {
		case <-r.Context().Done():
			return
		case evt, open := <-eventsCh:
			if !open {
				return
			}
			// ServiceImpl resolves event visibility once for all subscribers. Keep the detail
			// check for test doubles and legacy events that carry no resolved audience.
			if _, resolved := evt.canDeliver(currentUser.ID); !resolved {
				if _, err := h.service.GetIssueByID(r.Context(), evt.IssueID); err != nil {
					continue
				}
			}
			data, err := json.Marshal(evt)
			if err != nil {
				continue
			}
			if _, writeErr := fmt.Fprintf(w, "event: issue\ndata: %s\n\n", data); writeErr != nil {
				return
			}
			flusher.Flush()
		case <-ticker.C:
			// Periodic comment keepalive to prevent proxy idle connection drop
			if _, pingErr := fmt.Fprintf(w, ": keepalive\n\n"); pingErr != nil {
				return
			}
			flusher.Flush()
		}
	}
}