# openapi.yaml
openapi: 3.1.0
info:
  title: 6S Issue Management & Visual Reporting System API
  version: 1.0.0
  description: API phục vụ ghi nhận, đồng bộ offline, giải quyết và tính điểm sức khỏe 6S trong môi trường nhà xưởng.
servers:
  - url: /api
    description: Runtime-configured API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: ISSUE_CONFLICT
        message:
          type: string
          example: Phiên bản issue đã bị thay đổi bởi người dùng khác
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    TagDetail:
      type: object
      required:
        - code
        - name_vi
        - name_zh
        - name_en
        - category
        - status
      properties:
        code:
          type: string
          example: oil_leak
        name_vi:
          type: string
          example: Rò rỉ dầu
        name_zh:
          type: string
          example: 漏油
        name_en:
          type: string
          example: Oil leak
        category:
          type: string
          enum: [1S, 2S, 3S, 4S, 5S, 6S]
          example: 3S
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, MERGED]
          example: APPROVED
    Issue:
      type: object
      required:
        - id
        - client_uuid
        - version
        - creator_id
        - asset_id
        - assigned_team_id
        - assignee_id
        - cause_team_id
        - cause_status
        - category
        - location_code
        - tags
        - photo_before
        - status
        - created_at
      properties:
        id:
          type: integer
          example: 123
        client_uuid:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        version:
          type: integer
          example: 1
        creator_id:
          type: integer
          example: 12
        resolver_id:
          type: integer
          nullable: true
          example: 15
        asset_id:
          type: [integer, 'null']
          description: Luôn có mặt; null khi issue không gắn tài sản
        assigned_team_id:
          type: [integer, 'null']
          description: Team xử lý; luôn có mặt, null khi chưa phân công
        assignee_id:
          type: [integer, 'null']
          description: Người xử lý trực tiếp; luôn có mặt, null khi chưa phân công
        cause_team_id:
          type: [integer, 'null']
          description: Team chịu trách nhiệm nguyên nhân; luôn có mặt, không tự suy ra từ team xử lý
        cause_status:
          type: string
          enum: [UNVERIFIED, CONFIRMED, NOT_APPLICABLE]
          description: Trạng thái xác minh nguyên nhân; luôn có mặt, độc lập với cause_type
        responsibility_history:
          type: array
          description: Chỉ có ở GET /api/issues/{id}
          items:
            $ref: '#/components/schemas/ResponsibilityHistoryEntry'
        allowed_actions:
          description: Chỉ có ở GET /api/issues/{id}
          $ref: '#/components/schemas/IssueAllowedActions'
        category:
          type: string
          enum: [1S, 2S, 3S, 4S, 5S, 6S]
          example: 6S
        location_code:
          type: string
          example: LINE_A1
        location_name_vi_snapshot:
          type: string
          nullable: true
          description: Location name in Vietnamese captured with the issue, when available.
        location_name_zh_snapshot:
          type: string
          nullable: true
          description: Location name in Chinese captured with the issue, when available.
        location_name_en_snapshot:
          type: string
          nullable: true
          description: Location name in English captured with the issue, when available.
        location_snapshot_source:
          type: string
          nullable: true
          enum: [CLIENT_CAPTURE, SERVER_CAPTURE]
          description: Provenance of the display snapshot; never used for authorization.
        location_snapshot_recorded_at:
          type: string
          format: date-time
          nullable: true
          description: Server receipt time at which the snapshot was persisted.
        tags:
          type: array
          items:
            type: string
          example: ["oil_leak", "safety_gear"]
        tag_details:
          type: array
          description: Tag metadata and lifecycle status chips for all viewers
          items:
            $ref: '#/components/schemas/TagDetail'
        description:
          type: string
          nullable: true
          example: Dầu máy chảy ra sàn gần công tắc khẩn cấp
        reject_reason:
          type: string
          nullable: true
          example: Chưa dọn sạch dầu / Dây điện vẫn chưa bọc kỹ
        photo_before:
          type: string
          example: /uploads/before/3fa85f64-5717-4562-b3fc-2c963f66afa6_wide.jpg
        photo_detail:
          type: string
          nullable: true
          example: /uploads/before/3fa85f64-5717-4562-b3fc-2c963f66afa6_detail.jpg
        photo_after:
          type: string
          nullable: true
          example: /uploads/after/7ca85f64-5717-4562-b3fc-2c963f66af99.jpg
        score_rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        status:
          type: string
          enum: [OPEN, PENDING_REVIEW, CLOSED, INVALID]
          example: OPEN
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
...
    ResponsibilityHistoryEntry:
      type: object
      required: [id, action, changed_by, changed_by_name, old_value, new_value, created_at]
      properties:
        id:
          type: integer
        action:
          type: string
          enum: [ASSIGN, TRANSFER, CAUSE_VERIFY, OTHER]
          description: ASSIGN khi có người/đơn vị nhận xử lý; TRANSFER khi đổi người/đơn vị; CAUSE_VERIFY khi xác minh nguyên nhân; OTHER cho thay đổi trách nhiệm khác (ví dụ chỉ đổi tài sản)
        changed_by:
          type: [integer, 'null']
        changed_by_name:
          type: [string, 'null']
        old_value:
          anyOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        new_value:
          anyOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        created_at:
          type: string
          format: date-time
    IssueAllowedActions:
      type: object
      required: [assign, verify_cause, resolve, close]
      properties:
        assign:
          type: boolean
        verify_cause:
          type: boolean
        resolve:
          type: boolean
        close:
          type: boolean
...
security:
  - BearerAuth: []
paths:
...
  /issues/{id}/resolve:
    post:
      operationId: resolveIssue
      summary: Upload ảnh khắc phục (hỗ trợ offline sync với expected_version)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - resolved_client_uuid
                - expected_version
                - photo_after
              properties:
                resolved_client_uuid:
                  type: string
                  format: uuid
                expected_version:
                  type: integer
                  example: 1
                photo_after:
                  type: string
                  format: binary
      responses:
        '200':
          description: Khắc phục thành công, chuyển sang PENDING_REVIEW
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Issue'
        '409':
          description: Xung đột trạng thái hoặc phiên bản (Optimistic Locking)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /issues/{id}/close:
    post:
      operationId: closeIssue
      summary: Duyệt hoàn thành issue và chấm điểm Kaizen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score_rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 3
      responses:
        '200':
          description: Đóng issue thành công, cộng điểm Creator
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Issue'
        '403':
          description: Không đủ thẩm quyền đóng issue (ví dụ công nhân không được duyệt 6S Safety)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /issues/{id}/reopen:
    post:
      operationId: reopenIssue
      summary: Từ chối duyệt khắc phục, yêu cầu làm lại (Reopen)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reject_reason
              properties:
                reject_reason:
                  type: string
                  example: Chưa dọn sạch dầu / Dây điện vẫn chưa bọc kỹ
      responses:
        '200':
          description: Reopen thành công, trừ điểm chuyền và bắn thông báo khẩn
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Issue'
        '403':
          description: Không có quyền thực hiện
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /issues/{id}/invalid:
    post:
      operationId: invalidateIssue
      summary: Bác bỏ issue do báo sai hoặc spam (Requires permission `issue:invalidate`)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  example: Ảnh chụp không rõ / Báo sai thực tế
      responses:
        '200':
          description: Bác bỏ thành công, trừ điểm phạt Creator
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: '#/components/schemas/Issue'
        '403':
          description: Requires permission `issue:invalidate`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...